Updates project to use Android Gradle plugin v3.x
Updated the project to use the latest Gradle plugin and also bumped up all the other dependencies.

Also, the response builder in OkHttp 3.8 has made the message mandatory which Barricade was not passing, causing IllegalStateExceptions. I've fixed that issue by hardcoding a static message for now. In the future, we could make even the message configurable like we do for status codes.